#529353#精益看板

+++ agile-server/src/main/java/com/yusys/agile/set/stage/service/IStageService.java
@@ -22,7 +22,7 @@ public interface IStageService {
      * @param teamId
      * @return
      */
-    List<StageInstance> getStages(Integer stageType,Long teamId,Byte taskType) throws Exception;
+    List<StageInstance> getStages(Integer stageType,Long teamId,Integer taskType) throws Exception;
 
     /**
      * 敏捷看板根据工作项类型获取阶段信息
